changed error message and test_ name
